client/requirements.txt
PyQt6>=6.6.0
requests>=2.32.0
websocket-client>=1.8.0
